kexec-elf-rel-ppc64: Fix cast from pointer warning
Fixes warnings like these when building kexec for powerpc (32 bit):

  kexec-elf-rel-ppc64.c: warning: cast from pointer to integer of different size

+++ kexec/arch/ppc64/kexec-elf-rel-ppc64.c
@@ -104,7 +104,7 @@ void machine_apply_elf_rel(struct mem_ehdr *ehdr, struct mem_sym *sym,
 		break;
 
 	case R_PPC64_REL32:
-		*(uint32_t *)location = value - (uint32_t)(uint64_t)location;
+		*(uint32_t *)location = value - (uint32_t)location;
 		break;
 
 	case R_PPC64_TOC:
